renamed test util for jest sample test setup
